requirements_pip.txt
gunicorn==20.1.0

djangorestframework==3.13.1
asgiref==3.4.1
django==3.2.12
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.1.1

jsonformatter==0.3.1

pyunicore==0.9.15
PyJWT==2.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
requests==2.27.1
urllib3==1.26.8

psycopg2-binary==2.9.3
